Include building the C demo in the build.zig
Zig can also build C applications, so why not?

Additionally, the two demos for Zig & C are not the same; The C demo
contains an image, whereas the Zig one does not. It's clear the Zig one
was going to contain an image, but it was never actually added? So I
went ahead and added it.